+++ SSYPoller.h
@@ -23,7 +23,7 @@
 */
 + (BOOL)waitUntilInvocation:(NSInvocation*)invocation
 			 initialBackoff:(NSTimeInterval)initialBackoff
-			  backoffFactor:(float)backoffFactor
+			  backoffFactor:(CGFloat)backoffFactor
 				 maxBackoff:(NSTimeInterval)maxBackoff
 					timeout:(NSTimeInterval)timeout
 				  timeLimit:(NSTimeInterval)timeLimit
